fix(pkgs/zkm): Build internal go library
Recently ZKMips copied a Go library from their SDK to the ZKM repo.
Before we evaded the issue by not building the Rust package, which
builds the Go library.

This is no longer viable, because the library has been officially moved
(i.e. removed from the SDK). Now we're building it ourselves, and
removing the go build command from the appropriate build.rs